requirements.txt
streamlit==1.28.0
torch==2.5.1
torchvision==0.15.2
Pillow==10.0.0
opencv-python==4.8.0.74
numpy==1.24.3